requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==1.10.15
PyPDF2==3.0.1
python-dotenv==1.0.1
langchain==0.1.13
langchain-community==0.0.29
langchain-openai==0.0.8
faiss-cpu==1.7.4
